services/fanout-worker/requirements.txt
aiokafka==0.11.0
aiomysql==0.2.0
redis[asyncio]==5.2.0
pydantic-settings==2.6.1
opentelemetry-sdk==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
